+++ podaac/subsetter/time_converting.py
@@ -45,7 +45,10 @@ def convert_to_datetime(dataset: xr.Dataset, time_vars: list, instrument_type: s
         if np.issubdtype(dataset[var].dtype, np.dtype(float)) or np.issubdtype(dataset[var].dtype, np.float32):
             # adjust the time values from the start date
             if start_date:
-                dataset[var].values = [start_date + datetime.timedelta(seconds=i) for i in dataset[var].values]
+                # create array of the start time in datetime format
+                date_time_array = np.full(dataset[var].shape, start_date)
+                # add seconds since the start time to the start time to get the time at the data point
+                dataset[var].values = date_time_array.astype("datetime64[ns]") + dataset[var].astype('timedelta64[s]').values
                 continue
             # if there isn't a start_date, get it from the UTC variable
             utc_var_name = subset.compute_utc_name(dataset)